chore(tools): Fix submodule initialization check
Why
The previous check for missing submodules was incorrect.
It was checking for the existence of `.gitmodules` in the same directory
as init.bat, but the real path is one level up relative to that.
That prevented a 'git submodule update --init', which is necessary part
of the initialization process.

What
- Changed condition for submodules check to the right parent path.
- Removed unnecessary quoting in submodule error echo to be coherent
  with other error echos.