feat(bin): Add `get_full_package_name` command
Why:
We need a way to get the full package name, including version and extension,
for a given package.
This is useful in scripts to work with individual package files.
The existing `get_package_name` only returns the base name without version and
extension, which makes further processing reliant on additional search logic
that may produce unexpected complexity and edge cases.

What:
- Added `get_full_package_name` function.
- Added `gfpn` alias in `.env`.
- Added `search_full_package_name_in_folder` helper function.
- Modified `get_package_name` to use output of `current_tool` command.