fix(setup): setup_packages.sh fix package size check
Why
The package size check was failing due to a typo in the command used to obtain
the file size.
It was using `stats` instead of the correct `stat` command.
This prevented the size check from working correctly and could lead to not
processing files that were of normal size.

What
- Replaced `stats` with `stat` in the `package_file_size` calculation.