feat(install): Add ic alias, CPLX_INSTALL_COPY_ONLY
Why:
The install script copies files then executes itself remotely.
It was difficult to debug the copy part independently of the remote execution
part.
This change allows to stop the install after the copy to check what was
effectively copied.
A new `ic` alias makes it easy to launch the install in copy-only mode.

What:
- Added `CPLX_INSTALL_COPY_ONLY` environment variable to stop after copy.
- Added conditional code to check `CPLX_INSTALL_COPY_ONLY`.
- Added error handling for the scp system call.
- Added `ic` alias in `senv.bat`.