fix(env): Correctly handle return code of check_ldd
Why:
The `mirror_system_executable` function was not correctly handling the return
code of the `check_ldd` function.
It was using `if check_ldd ...; then ret=1; fi`, meaning that the error code
would always be set.
This was changed so the error code is set only when ldd check failed.
Also, an explicit return 0 was added to the `_copy_element` function that always
returns 0 when ends without error.
Before this change, the return code would depend upon the previous commands,
often leading to misleading and unnecessary errors.

What:
- Inverted the condition in the `if` statement in the copy section to ensure the
  return code is set correctly in case of error
- Inverted the condition on the `if` statement to correct error handling in the
  regular file check.
  In this case the return code was inverted because the if is negated
- Added `warning` messages to provide more context in case of errors after
  `ldd check`
- Added `return 0` to `_copy_element` function at the very end