feat(install): Add CPLX_CONFIG_DONE property
Why:
The previous implementation relied on a hardcoded check for "`creating Makefile`"
in `config.log` to determine if the configure step had completed successfully.
This approach is not robust and can lead to false positives or negatives.
This change introduces a new property `CPLX_CONFIG_DONE`, allowing users to
specify the value that should be grepped and the file to grep, to confidently
signal successful configuration.

What:
- Added `CPLX_CONFIG_DONE` property to `setup.sh` and `add_tool.bat`
- Modified install script to use `CPLX_CONFIG_DONE` to determine if
  configuration was successful.
- Updated `senv.local.tpl` with default values for `CPLX_CONFIG_DONE` for
  existing tools.
- Added logic to handle the '`default`' value for `CPLX_CONFIG_DONE`.
- Added validation and error handling for `CPLX_CONFIG_DONE`.
- Added documentation for the new property.