fix(install): handle built-in packages in pkg-config
Why: The pkg-config files generated by some tools include absolute paths, which
are not compatible with the current tool sysroot.
This change ensures that the paths in the pkg-config files are relative to the
sysroot directory of the current tool when the package is updated.

What:
- Added a new conditional branch to handle paths like
`/home/gitea2/cplx/tools/openssl111/openssl111-1.1.1w/usr`  by rewriting them as
relative paths to the current tool sysroot directory.
- Preserved the existing logic for other path formats for backwards
  compatibility.
- Added a fallback to handle cases where the new regex does not match correctly.