feat(libgpg-error): Add libgpg-error service
Why:
The project requires libgpg-error for cryptographic operations.
This change adds libgpg-error as a service to the project, enabling its
installation and usage within the development environment.
This enhances security and compatibility with other cryptographic libraries.

But... autopoint does not work, needs gettext-tools, which does not exist on RHEL
So stop for now: not compiled

What:
- Added `libgpg-error` to the list of services and tools to recompile in
  `setup.tpl.properties` and `cplx.tpl.properties`.
- Created `libgpg-error_install_functions.sh` containing installation functions
  for libgpg-error.
- Added configuration for libgpg-error in `senv.local.tpl`.
- Created package lists for libgpg-error in `libgpg-error_centos_8_x86_64.txt`
  and `libgpg-error_rhel_7.9_x86_64.txt`.
- Added required packages for libgpg-error compilation to
  `packages_rhel_7.9_x86_64.txt`.