fix(glibc): Correct glibc build process
Why:
The glibc build process was incorrect and failed.
It tried to build within the source directory instead of using a dedicated build
directory, which is mandatory for glibc (out-of-tree build).
It also assumed hardcoded path to `gcc`, which is not suitable for
cross-compilation.
Moreover, the build directory was not created in predictable location relative
to the source.

What:
- Added `glibc_setenv` to set `COMPILER_PATH` and augment `PATH`
- Set object directory location predictably relative to source to
  `${tool_src}/../build`
- Modified `configure` and `build` functions to build in the object directory
- Removed the faulty `sed` command in `configure` aimed to update `configure`
- Replaced hardcoded checks for gcc with dynamic approach
- Updated the Makefile to reflect the changes and avoid make errors
- Added checks for `CPLX_CHECK_SRC` variable existence and made build
  conditional to presence of `CPLX_CHECK_SRC`
- The execution location for make changed from source directory to the build
  directory.
- Updated log and error messages for build failures.