feat(install): Introduce glibc wrapper scripts
Why:
To ensure that executables within the glibc target folder use the correct
versions of libraries.
This avoids potential conflicts with system libraries and provides a consistent
execution environment.
This also permits to use the development libraries instead of the operating
system ones if they do not exist.
Usually happens when the operating system is too old, or when a too recent
compiler is used to build the binaries.

What:
- Added `create_wrappers.py` to generate wrapper scripts for ELF executables.
- Added `setenv` script to set required environment variables, including library
  paths.
- Wrapper scripts use `NEW_LDSO` and `NEW_LIB_DIR` to execute binaries with the
  correct dynamic linker and library paths.
- Implemented `update_path_variable` function to safely modify `PATH` like
  variables (`MAN_PATH`, `PERLLIB`, `LD_LIBRARY_PATH`), preventing duplicate
  entries.
- Created symbolic links from `tool_dir/current`  to `../current` and
  `tool_dir/root` to `../root`.
- Added alias for `man` to use specific wrapper `manw.sh`.