chore(pkgs): Add .keep file for package directory
Why:
The `.keep` file helps make sure the `src/install/env/pkgs/` directory stays in
its place in the project's setup.
This is good because even though packages themselves go into a central
`tools/pkgs/` spot, this directory is still part of how things are expected to
be put together.
It also clears up where packages for this specific part of the system are
fetched or put, which means less head-scratching later.

What:
- Put a `.keep` file in the `src/install/env/pkgs/` folder.
- The file now spells out that packages are copied to a central spot.